Optimized rerender with redux
